Добавлен сценарий регистрации
