+++ templates/guest.php
@@ -7,5 +7,5 @@
         <p>После создания аккаунта, пользователь может начать вносить свои дела, деля их по проектам и указывая сроки.</p>
     </div>
 
-    <a class="welcome__button button" href="index.php?register">Зарегистрироваться</a>
+    <a class="welcome__button button" href="register.php">Зарегистрироваться</a>
 </section>
